restructure facade pattern
